Add modular analysis components and DI
